+++ exporters/otlp/otlp_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package otlp
+package otlp_test
 
 import (
 	"context"
@@ -20,25 +20,103 @@ import (
 	"sync"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"go.opentelemetry.io/otel/exporters/otlp"
+	metricpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/metrics/v1"
+	tracepb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/trace/v1"
+	"go.opentelemetry.io/otel/exporters/otlp/internal/transform"
+
+	metricsdk "go.opentelemetry.io/otel/sdk/export/metric"
+	tracesdk "go.opentelemetry.io/otel/sdk/export/trace"
 )
+
+type stubProtocolDriver struct {
+	started         int
+	stopped         int
+	tracesExported  int
+	metricsExported int
+
+	injectedStartError error
+	injectedStopError  error
+
+	rm []metricpb.ResourceMetrics
+	rs []tracepb.ResourceSpans
+}
+
+var _ otlp.ProtocolDriver = (*stubProtocolDriver)(nil)
+
+func (m *stubProtocolDriver) Start(ctx context.Context) error {
+	m.started++
+	select {
+	case <-ctx.Done():
+		return ctx.Err()
+	default:
+		return m.injectedStartError
+	}
+}
+
+func (m *stubProtocolDriver) Stop(ctx context.Context) error {
+	m.stopped++
+	select {
+	case <-ctx.Done():
+		return ctx.Err()
+	default:
+		return m.injectedStopError
+	}
+}
+
+func (m *stubProtocolDriver) ExportMetrics(parent context.Context, cps metricsdk.CheckpointSet, selector metricsdk.ExportKindSelector) error {
+	m.metricsExported++
+	rms, err := transform.CheckpointSet(parent, selector, cps, 1)
+	if err != nil {
+		return err
+	}
+	for _, rm := range rms {
+		if rm == nil {
+			continue
+		}
+		m.rm = append(m.rm, *rm)
+	}
+	return nil
+}
+
+func (m *stubProtocolDriver) ExportTraces(ctx context.Context, ss []*tracesdk.SpanSnapshot) error {
+	m.tracesExported++
+	for _, rs := range transform.SpanData(ss) {
+		if rs == nil {
+			continue
+		}
+		m.rs = append(m.rs, *rs)
+	}
+	return nil
+}
+
+func (m *stubProtocolDriver) Reset() {
+	m.rm = nil
+	m.rs = nil
+}
+
+func newExporter(t *testing.T, opts ...otlp.ExporterOption) (*otlp.Exporter, *stubProtocolDriver) {
+	driver := &stubProtocolDriver{}
+	exp, err := otlp.NewExporter(context.Background(), driver, opts...)
+	require.NoError(t, err)
+	return exp, driver
+}
 
 func TestExporterShutdownHonorsTimeout(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Minute)
 	defer cancel()
 
-	e := NewUnstartedExporter()
-	orig := e.cc.closeBackgroundConnectionDoneCh
-	e.cc.closeBackgroundConnectionDoneCh = func(ch chan struct{}) {
-		go func() {
-			<-ctx.Done()
-			orig(ch)
-		}()
-	}
+	e := otlp.NewUnstartedExporter(&stubProtocolDriver{})
 	if err := e.Start(ctx); err != nil {
 		t.Fatalf("failed to start exporter: %v", err)
 	}
 
 	innerCtx, innerCancel := context.WithTimeout(ctx, time.Microsecond)
+	<-time.After(time.Second)
 	if err := e.Shutdown(innerCtx); err == nil {
 		t.Error("expected context DeadlineExceeded error, got nil")
 	} else if !errors.Is(err, context.DeadlineExceeded) {
@@ -51,14 +129,7 @@ func TestExporterShutdownHonorsCancel(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Minute)
 	defer cancel()
 
-	e := NewUnstartedExporter()
-	orig := e.cc.closeBackgroundConnectionDoneCh
-	e.cc.closeBackgroundConnectionDoneCh = func(ch chan struct{}) {
-		go func() {
-			<-ctx.Done()
-			orig(ch)
-		}()
-	}
+	e := otlp.NewUnstartedExporter(&stubProtocolDriver{})
 	if err := e.Start(ctx); err != nil {
 		t.Fatalf("failed to start exporter: %v", err)
 	}
@@ -77,7 +148,7 @@ func TestExporterShutdownNoError(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Minute)
 	defer cancel()
 
-	e := NewUnstartedExporter()
+	e := otlp.NewUnstartedExporter(&stubProtocolDriver{})
 	if err := e.Start(ctx); err != nil {
 		t.Fatalf("failed to start exporter: %v", err)
 	}
@@ -89,7 +160,7 @@ func TestExporterShutdownNoError(t *testing.T) {
 
 func TestExporterShutdownManyTimes(t *testing.T) {
 	ctx := context.Background()
-	e, err := NewExporter(ctx)
+	e, err := otlp.NewExporter(ctx, &stubProtocolDriver{})
 	if err != nil {
 		t.Fatalf("failed to start an exporter: %v", err)
 	}
@@ -110,6 +181,99 @@ func TestExporterShutdownManyTimes(t *testing.T) {
 	for _, err := range errs {
 		if err != nil {
 			t.Fatalf("failed to shutdown exporter: %v", err)
+		}
+	}
+}
+
+func TestSplitDriver(t *testing.T) {
+	driverTraces := &stubProtocolDriver{}
+	driverMetrics := &stubProtocolDriver{}
+	config := otlp.SplitConfig{
+		ForMetrics: driverMetrics,
+		ForTraces:  driverTraces,
+	}
+	driver := otlp.NewSplitDriver(config)
+	ctx := context.Background()
+	assert.NoError(t, driver.Start(ctx))
+	assert.Equal(t, 1, driverTraces.started)
+	assert.Equal(t, 1, driverMetrics.started)
+	assert.Equal(t, 0, driverTraces.stopped)
+	assert.Equal(t, 0, driverMetrics.stopped)
+	assert.Equal(t, 0, driverTraces.tracesExported)
+	assert.Equal(t, 0, driverTraces.metricsExported)
+	assert.Equal(t, 0, driverMetrics.tracesExported)
+	assert.Equal(t, 0, driverMetrics.metricsExported)
+
+	assert.NoError(t, driver.ExportMetrics(ctx, discCheckpointSet{}, metricsdk.StatelessExportKindSelector()))
+	assert.NoError(t, driver.ExportTraces(ctx, []*tracesdk.SpanSnapshot{discSpanSnapshot()}))
+	assert.Len(t, driverTraces.rm, 0)
+	assert.Len(t, driverTraces.rs, 1)
+	assert.Len(t, driverMetrics.rm, 1)
+	assert.Len(t, driverMetrics.rs, 0)
+	assert.Equal(t, 1, driverTraces.tracesExported)
+	assert.Equal(t, 0, driverTraces.metricsExported)
+	assert.Equal(t, 0, driverMetrics.tracesExported)
+	assert.Equal(t, 1, driverMetrics.metricsExported)
+
+	assert.NoError(t, driver.Stop(ctx))
+	assert.Equal(t, 1, driverTraces.started)
+	assert.Equal(t, 1, driverMetrics.started)
+	assert.Equal(t, 1, driverTraces.stopped)
+	assert.Equal(t, 1, driverMetrics.stopped)
+	assert.Equal(t, 1, driverTraces.tracesExported)
+	assert.Equal(t, 0, driverTraces.metricsExported)
+	assert.Equal(t, 0, driverMetrics.tracesExported)
+	assert.Equal(t, 1, driverMetrics.metricsExported)
+}
+
+func TestSplitDriverFail(t *testing.T) {
+	ctx := context.Background()
+	for i := 0; i < 16; i++ {
+		var (
+			errStartMetric error
+			errStartTrace  error
+			errStopMetric  error
+			errStopTrace   error
+		)
+		if (i & 1) != 0 {
+			errStartTrace = errors.New("trace start failed")
+		}
+		if (i & 2) != 0 {
+			errStopTrace = errors.New("trace stop failed")
+		}
+		if (i & 4) != 0 {
+			errStartMetric = errors.New("metric start failed")
+		}
+		if (i & 8) != 0 {
+			errStopMetric = errors.New("metric stop failed")
+		}
+		shouldStartFail := errStartTrace != nil || errStartMetric != nil
+		shouldStopFail := errStopTrace != nil || errStopMetric != nil
+
+		driverTraces := &stubProtocolDriver{
+			injectedStartError: errStartTrace,
+			injectedStopError:  errStopTrace,
+		}
+		driverMetrics := &stubProtocolDriver{
+			injectedStartError: errStartMetric,
+			injectedStopError:  errStopMetric,
+		}
+		config := otlp.SplitConfig{
+			ForMetrics: driverMetrics,
+			ForTraces:  driverTraces,
+		}
+		driver := otlp.NewSplitDriver(config)
+		errStart := driver.Start(ctx)
+		if shouldStartFail {
+			assert.Error(t, errStart)
+		} else {
+			assert.NoError(t, errStart)
+		}
+		errStop := driver.Stop(ctx)
+		if shouldStopFail {
+			assert.Error(t, errStop)
+		} else {
+			assert.NoError(t, errStop)
 		}
 	}
 }